requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# LLM Integration
google-generativeai==0.3.2
pydantic-settings==2.0.3

# Data Validation
pydantic==2.5.0

# HTTP Client
httpx==0.25.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: For system control actions
pywin32==306; sys_platform == "win32"
